src/main/java/com/jmp/controller/MenteeController.java
package com.jmp.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import com.jmp.dto.MenteeDto;
import com.jmp.service.MenteeService;

@RestController
@RequestMapping("/mentees")
public class MenteeController {

    @Autowired
    private MenteeService menteeService;

    @ResponseStatus(HttpStatus.OK)
    @RequestMapping(path = "/{id}", method = RequestMethod.GET)
    public MenteeDto find(@PathVariable("id") int id) {
        return menteeService.find(id);
    }

    @ResponseStatus(HttpStatus.CREATED)
    @RequestMapping(method = RequestMethod.POST)
    public void create(@RequestBody MenteeDto mentorDto) {
        menteeService.create(mentorDto);
    }

    @ResponseStatus(HttpStatus.OK)
    @RequestMapping(method = RequestMethod.PUT)
    public void update(@RequestBody MenteeDto mentorDto) {
        menteeService.update(mentorDto);
    }

    @ResponseStatus(HttpStatus.NO_CONTENT)
    @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
    public void delete(@PathVariable("id") int id) {
        menteeService.delete(id);
    }
}
